requirements.txt
anthropic>=0.49.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
rich>=13.0.0
httpx>=0.27.0
pyngrok>=7.0.0
